ogsmd: Split the info for concatenated short messages in multiple keys
There's now csm_id, csm_num and csm_seq for the id of the csm, the
total number of messages for that csm and the sequence of this message.
The features will only have simple types as values (no lists or maps)
to make life easier especially for the C guys.
WARNING: The final names for the features are not decided upon yet and
are bound to change!