ogsmd: sms.py: Don't encode the last 7 bit if they are padding
This fixes a bug where the end of the message had a '@' character
appended to it. This would always happen when the message was of such a
length that the last bit of the character was encoded in a new byte. The
remaining 7bit would then also get decoded (all zero -> @).