opimd: Created 'Backend' and 'Domain' classes
It has to be done because we don't want to register the backends and the domains at import time.
So instead we register only the classes at import time using the metaclasses, and then we can register all the
domains and the backends.
I also renamed the modules to remove the '-' characters, cause we have to import the modules manually.