oeventsd: using yaml CLoader
This make things faster, but there is a problem : it is not possible to tell yaml to automatically
interpret the functions anymore, so we have to do it manually after each call to yaml.load, and iterate the structure
looking for functions. It is really not a very good solution.