ogpsd: msm.py: activate/desactivate gps by launching/killing the "gps" program
I used subprocess for launching and killing the gps program.

Here are the limitations:
* playing too fast with gps's on/off in shr-settings can crash
  (and so reboot) the phone
* The gps program must not be launched manually(not within frameworkd)

Note that if we use shell=True for launching the gps program,
the .kill() kills the shell instead of the gps program.
Thanks a lot Yhg1s in #python on Freenode for help on debugging that

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@no-log.org>