Make blink rate for Oeventsd configurable
Allow 2 optional parameters to oevents blink statements - durationOn and
durationOff.  The default is the previous blink rate.

GTA04 hardware supports fixed blink rates.  If the requested blink rate
is not close (within 12%) of a supported rate, blinking will not
continue in suspend.  Set the blink rate to the closest supported rate.

Signed-off-by: Ben Deering <ben_deering@swissmail.org>
Signed-off-by: Lukas Märdian <luk@slyon.de>